Add files via upload
In this project we are going to implement, Round Robin Scheduling Algorithm, it is one of the oldest, simplest, and fairest and most widely used scheduling algorithms, designed especially for time-sharing systems. A small unit of time, called time-slice or quantum, is defined. All runnable processes are kept in a circular queue. The CPU scheduler goes around this queue, allocating the CPU to each process for a time. New processes are added to the tail of the queue with a limited slice of time. Where each process gets an equal share of something in turns, handling all processes without priority. Which is mostly used for multitasking. Round-robin scheduling can also be applied to other scheduling problems, such as data packet scheduling in computer networks. Round robin is most efficiently used method in operating system. The programme uses the OpenGL API to render graphics. Hence it is basically a C or C++ Computer Graphics Program using OpenGL.